+++ src/index.cpp
@@ -85,10 +85,10 @@ void index<ColorClasses>::build_from(ccdbg_builder const& builder) {
             [&](ggcat::Slice<char> const unitig, ggcat::Slice<uint32_t> const colors,
                 bool /* same_color */) {
                 auto lookup_result = m_k2u.lookup_advanced(unitig.data);
-                uint32_t unitig_id = lookup_result.contig_id;
+                uint64_t unitig_id = lookup_result.contig_id;
                 uint32_t color_id = u2c(unitig_id);
                 for (uint64_t i = 1; i != unitig.size - m_k2u.k() + 1; ++i) {
-                    uint32_t got = m_k2u.lookup_advanced(unitig.data + i).contig_id;
+                    uint64_t got = m_k2u.lookup_advanced(unitig.data + i).contig_id;
                     if (got != unitig_id) {
                         std::cout << "got unitig_id " << got << " but expected " << unitig_id
                                   << std::endl;
